refactor: move Orderbook to distinct dir
