+++ src/App.tsx
@@ -19,10 +19,10 @@ import CreateWallet from '@/components/CreateWallet'
 import JamLanding from '@/components/JamLanding'
 import LoginPage from '@/components/Login'
 import { Logs } from '@/components/Logs'
-import { Orderbook } from '@/components/Orderbook'
 import SwitchWallet from '@/components/SwitchWallet'
 import { EarnPage } from '@/components/earn/EarnPage'
 import { Layout } from '@/components/layout/Layout'
+import { Orderbook } from '@/components/orderbook/Orderbook'
 import { Receive } from '@/components/receive/Receive'
 import { SendPage } from '@/components/send/SendPage'
 import { RescanChain } from '@/components/settings/RescanChain'
